chore(test): move tls server tests into main test file
